refactor(sys)!: remove foreign symbols from bindings
Leave only NGINX code and referenced types.

Unfiltered generation of bindings for all the system includes and
dependencies slows down the build and causes lots of odd issues
(#322, rust-bindgen#3498).

Most of the removed symbols have replacements in the `openssl-sys` and
`libc` crates.  Note that the remaining local types are not compatible
with the equivalent types from external crates, and some unsafe pointer
casts may be necessary.